+++ src/wallet/config.ts
@@ -42,7 +42,7 @@ export class Config {
     this.state_entity_endpoint = DEFAULT_STATE_ENTITY_ENPOINT;
     this.swap_conductor_endpoint = DEFAULT_STATE_ENTITY_ENPOINT;
     this.electrum_config = {
-      host: '0.0.0.0',
+      host: 'wallet.mercurywallet.io',
       port: 50004,
       protocol: 'wss',
     }
