+++ src/wallet/mocks/mock_http_client.ts
@@ -78,13 +78,13 @@ export const ROOT_INFO = {
 
 export const STATECHAIN_INFOS = [{
   utxo: "f62c9b74e276843a5d0fe0d3d0f3d73c06e118b822772c024aac3d840fbad3ce:1",
-  amount: 100,
+  amount: 1000,
   chain: [{ data: "03ffac3c7d7db6308816e8589af9d6e9e724eb0ca81a44456fef02c79cba984477", next_state: null }],
   locktime: 1000
 },
 {
   utxo: "f62c9b74e276843a5d0fe0d3d0f3d73c06e118b822772c024aac3d840fbad3de:1",
-  amount: 100,
+  amount: 1000,
   chain: [{ data: "03ffac3c7d7db6308816e8589af9d6e9e724eb0ca81a44456fef02c79cba984478", next_state: null }],
   locktime: 1000
 }
@@ -94,7 +94,7 @@ export const STATECHAIN_INFOS = [{
 
 export const STATECHAIN_INFO =  {
   utxo: "0158f2978e5c2cf407970d7213f2b4289993b2fe3ef6aca531316cdcf347cc41:1",
-  amount: 100,
+  amount: 1000,
   chain: [{ data: "03ffac3c7d7db6308816e8589af9d6e9e724eb0ca81a44456fef02c79cba984477", next_state: null }],
   locktime: 1000
 }
@@ -119,7 +119,7 @@ export const STATECOIN_PROOF_KEY_DER =
 // Protocol run 1: transfer
 export const TRANSFER_SENDER = {x1:{secret_bytes:[4, 246, 192, 43, 69, 23, 178, 71, 110, 243, 18, 94, 63, 218, 196, 108, 233, 206, 106, 46, 41, 119, 58, 111, 27, 34, 131, 168, 191, 159, 116, 47, 105, 16, 0, 229, 91, 242, 129, 148, 9, 32, 169, 35, 147, 81, 145, 89, 120, 113, 52, 219, 120, 240, 95, 65, 18, 151, 19, 183, 175, 85, 35, 45, 4, 31, 193, 58, 52, 12, 133, 131, 108, 240, 63, 84, 93, 72, 57, 107, 174, 117, 10, 86, 11, 178, 235, 78, 195, 200, 49, 189, 224, 117, 14, 185, 5, 56, 84, 96, 142, 194, 95, 115, 175, 248, 103, 37, 171, 122, 249, 121, 183, 251, 135, 172, 10, 43, 14, 156, 21, 216, 155, 241, 67]},proof_key:"03ffac3c7d7db6308816e8589af9d6e9e724eb0ca81a44456fef02c79cba984477"};
 export const TRANSFER_MSG3 = {shared_key_id:"96c77c1a-b3c1-4e2a-9e86-636f59306bbc",t1:{secret_bytes:[4, 236, 178, 53, 9, 218, 194, 137, 0, 80, 132, 88, 82, 233, 158, 111, 162, 176, 232, 242, 83, 145, 84, 237, 210, 14, 216, 37, 106, 178, 113, 194, 216, 124, 152, 234, 104, 207, 84, 50, 69, 2, 141, 244, 153, 196, 12, 15, 221, 230, 202, 18, 138, 142, 233, 40, 105, 46, 26, 246, 75, 65, 16, 45, 195, 220, 181, 217, 215, 157, 150, 114, 135, 83, 46, 39, 152, 130, 187, 242, 104, 125, 236, 253, 222, 76, 97, 111, 114, 95, 147, 241, 98, 85, 228, 71, 243, 74, 168, 25, 24, 237, 123, 50, 251, 204, 219, 74, 169, 37, 155, 177, 122, 36, 23, 36, 157, 169, 111, 107, 168, 152, 138, 58, 221]},statechain_sig:{purpose:"TRANSFER",data:"028a9b66d0d2c6ef7ff44a103d44d4e9222b1fa2fd34cd5de29a54875c552abd41",sig:"3044022026a22bb2b8c0e43094d9baa9de1abd1de914b59f8bbcf5b740900180da575ed10220544e27e2861edf01b5c383fc90d8b1fd41211628516789f771b2c3536e650bda"},"statechain_id":"f7ac71c1-0937-4718-bc9b-7f4d77321981",tx_backup_psm:{shared_key_ids:["96c77c1a-b3c1-4e2a-9e86-636f59306bbc"],protocol:"Transfer",tx_hex:"02000000000101bb793d9d0a489b50c95ec03f147d647a926459e05301894a351f4684899d12a50000000000ffffffff02b802000000000000160014d19cb7bef37606bff26c988fc7986b6999412cde04000000000000001600141319a227287cfac4d8660830f4c9b0e1724a810001020000c55e1e00",input_addrs:["02303b83bc0a2980d353cc4a9474afb35e5d8d183b2ad9a1722e4a957fc15aa9b0"],input_amounts:[1000],proof_key:"03ffac3c7d7db6308816e8589af9d6e9e724eb0ca81a44456fef02c79cba984477"},rec_se_addr:{tx_backup_addr:"tb1q6xwt00hnwcrtlunvnz8u0xrtdxv5ztx7pj44cp",proof_key:"028a9b66d0d2c6ef7ff44a103d44d4e9222b1fa2fd34cd5de29a54875c552abd41"}}
-export const STATECHAIN_INFO_AFTER_TRANSFER = {utxo:"a5129d8984461f354a890153e05964927a647d143fc05ec9509b480a9d3d79bb:0",amount:1700,chain:[{data:"03ffac3c7d7db6308816e8589af9d6e9e724eb0ca81a44456fef02c79cba984477",next_state:null}],locktime:1990441}
+export const STATECHAIN_INFO_AFTER_TRANSFER = {utxo:"a5129d8984461f354a890153e05964927a647d143fc05ec9509b480a9d3d79bb:0",amount:1000,chain:[{data:"03ffac3c7d7db6308816e8589af9d6e9e724eb0ca81a44456fef02c79cba984477",next_state:null}],locktime:1990441}
 export const STATECOIN_PROOF_KEY_DER_AFTER_TRANSFER = {__D:[12,167,86,244,1,71,143,177,161,102,210,121,69,80,29,138,245,154,218,28,181,82,197,152,80,157,252,180,148,244,117,184],__Q:[2,138,155,102,208,210,198,239,127,244,74,16,61,68,212,233,34,43,31,162,253,52,205,93,226,154,84,135,92,85,42,189,65],chainCode:[17,103,219,181,100,221,223,210,34,16,230,138,129,23,181,247,233,17,25,193,254,149,5,150,20,118,66,82,146,33,29,195],network:{messagePrefix:"\u0018Bitcoin Signed Message:\n",bech32:"tb",bip32:{public:70617039,private:70615956},pubKeyHash:111,scriptHash:196,wif:239},__DEPTH:3,__INDEX:3,__PARENT_FINGERPRINT:278430330,lowR:false}
 export const TRANSFER_PUBKEY = {key:"03b1e51eb08539dcbf58995f9e75519071d4acd88d16314fd7231e707c75707098"}
 export const TRANSFER_RECEIVER = {new_shared_key_id:"d91553ca-8cab-4eef-b315-28583fd4180b",s2_pub:{x:"abc03bfe5e2fb3d54a58069bb06cef5d25533da24d928679b1fa9e9c97b6e0e0",y:"d47dc07001729dd5adb19c52a1ce1b9c360b0e15abb5dabd652dbba787fe35c0"}}
